Use messagepack extensions payloads that can be transmitted
As far as I have been able to understand, cl-messagepack's current
implementation of extensions doesn't allow you to encode arbitrary
data, associate it with some type, and then communicate this across a
network. cl-messagepack's implementation is more like you have some
instance of an object, you assign an ID to it, you can then
communicate this ID across the network and back, and then lookup this
object in a hashtable using the ID. This restricts you to keeping the
data in the same lisp image / memory.

This is a first stab at what I think is a "true" implementation of
messagepack's extensions. It is incomplete, but I'm making the PR for
visibility.

Use it to encode complex floats with the following:

```
(defun my-encode-complex (data stream)
  (encode-fixext-8
   (append
    (rest (coerce (flexi-streams:with-output-to-sequence (s) (encode-float (realpart data) s)) 'list))
    (rest (coerce (flexi-streams:with-output-to-sequence (s) (encode-float (imagpart data) s)) 'list)))
   stream
   0))

(defun my-decode-complex (len stream)
  (complex (sb-kernel:make-single-float (ub32->sb32 (load-big-endian stream 4)))
           (sb-kernel:make-single-float (ub32->sb32 (load-big-endian stream 4))))
  ;; (or #+sbcl (sb-kernel:make-single-float (ub32->sb32 (load-big-endian stream 4)))
  ;;     #-(or sbcl) (error "No floating point support yet."))
  ;; (or #+sbcl (sb-kernel:make-double-float (ub32->sb32 (load-big-endian stream 4))
  ;;                                         (load-big-endian stream 4))
  ;;     #+ccl (ccl::double-float-from-bits (load-big-endian stream 4)
  ;;                                        (load-big-endian stream 4))
  ;;     #-(or sbcl ccl) (error "No floating point support yet."))
  )

(register-extension-dispatcher 0 #'complexp #'my-encode-complex

CL-USER> (mpk::decode (mpk::encode (list #C(1.2 1.3)
                                         #C(0.0 1.0)
                                         (list #C(1.2 1.3)))))

```